refacturing Not Acks Scraper & Get Key Gen State for future uses.
